requirements.txt
discord.py
websockets
PyYAML